Move tests from PathUtil test to case sensitive classes
