+++ subprojects/snapshots/src/test/groovy/org/gradle/internal/snapshot/PathUtilTest.groovy
@@ -19,12 +19,7 @@ package org.gradle.internal.snapshot
 import spock.lang.Specification
 import spock.lang.Unroll
 
-import static org.gradle.internal.snapshot.PathUtil.compareChars
-import static org.gradle.internal.snapshot.PathUtil.compareCharsIgnoringCase
-import static org.gradle.internal.snapshot.PathUtil.equalChars
 import static org.gradle.internal.snapshot.PathUtil.getFileName
-import static org.gradle.internal.snapshot.PathUtil.isChildOfOrThis
-import static org.gradle.internal.snapshot.PathUtil.sizeOfCommonPrefix
 
 @Unroll
 class PathUtilTest extends Specification {
@@ -42,211 +37,4 @@ class PathUtilTest extends Specification {
         "C:${File.separator}Windows/system${File.separator}win32" | "win32"
     }
 
-    def "common prefix of #prefix and #path at #offset is #result"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert sizeOfCommonPrefix(prefix, path, offset, it) == result
-        }
-
-        where:
-        prefix       | path          | offset | result
-        '/root'      | '/'           | 0      | 0
-        '/root'      | '/root'       | 0      | 5
-        '/root/some' | '/root/other' | 0      | 5
-    }
-
-    def "can compare size of common prefix"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert Integer.signum(PathUtil.compareWithCommonPrefix(prefix, path, offset, it)) == result
-        }
-        if (result) {
-            assert Integer.signum(prefix <=> path.substring(offset)) == result
-        }
-
-        where:
-        prefix             | path                    | offset | result
-        "hello/world"      | "hello/other"           | 0      | 0
-        "hello/world"      | "/var/hello/other"      | 5      | 0
-        "hello/world"      | "/var/hello/world"      | 5      | 0
-        "hello/world"      | "/var/hello\\world"     | 5      | 0
-        "hello/world"      | "/var/hello/world/next" | 5      | 0
-        "hello/world"      | "/var/hello1/other"     | 5      | -1
-        "hello1/world"     | "/var/hello/other"      | 5      | 1
-        "hello/world/some" | "/var/hello/other"      | 5      | 0
-        "hello/world"      | "/var/hello1/other"     | 5      | -1
-        "bbc/some"         | "/var/abc/other"        | 5      | 1
-        "hello/world/some" | "/var/hello/other"      | 1      | -1
-        "hello/world/some" | "/var/hello/other"      | 5      | 0
-    }
-
-    def "size of common prefix of #prefix with #path at offset #offset is #result"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert sizeOfCommonPrefix(prefix, path, offset, it) == result
-        }
-
-        where:
-        prefix              | path                    | offset | result
-        "hello/world"       | "hello/other"           | 0      | 5
-        "hello/world"       | "/var/hello/other"      | 5      | 5
-        "hello/world"       | "/var/hello/world"      | 5      | 11
-        "hello/world"       | "/var/hello\\world"     | 5      | 11
-        "hello/world"       | "/var/hello/world/next" | 5      | 11
-        "hello/world"       | "/var/hello1/other"     | 5      | 0
-        "hello1/world"      | "/var/hello/other"      | 5      | 0
-        "hello/world/some"  | "/var/hello/other"      | 5      | 5
-        "hello/world"       | "/var/hello1/other"     | 5      | 0
-        "bbc/some"          | "/var/abc/other"        | 5      | 0
-        "/hello/world/some" | "/var/hello/other"      | 0      | 0
-        "/hello/world/some" | "/var/hello/other"      | 4      | 6
-    }
-
-    def "#prefix is child of or this of #path at offset #offset: #result"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert isChildOfOrThis(path, offset, prefix, it)
-            assert sizeOfCommonPrefix(prefix, path, offset, it) == prefix.length()
-        }
-
-        where:
-        prefix         | path                       | offset | result
-        "hello/world"  | "hello/world\\inside"      | 0      | true
-        "hello/world"  | "/var/hello/world\\inside" | 5      | true
-        "hello\\world" | "/var/hello/world\\inside" | 5      | true
-    }
-
-    def "separator is smaller than every other character"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert Integer.signum(PathUtil.compareWithCommonPrefix(prefix, path, offset, it)) == result
-        }
-
-        where:
-        prefix              | path                    | offset | result
-        "hello/world"       | "/var/hello-other"      | 5      | -1
-    }
-
-    def "can compare to child of this"() {
-        expect:
-        CaseSensitivity.values().each {
-            assert Integer.signum(PathUtil.compareToChildOfOrThis(prefix, path, offset, it)) == result
-        }
-        if (result) {
-            assert Integer.signum(prefix <=> path.substring(offset)) == result
-        }
-
-        where:
-        prefix              | path                    | offset | result
-        "hello/world"       | "hello/other"           | 0      | 1
-        "hello/world"       | "/var/hello/other"      | 5      | 1
-        "hello/world"       | "/var/hello/world"      | 5      | 0
-        "hello/world"       | "/var/hello/world/next" | 5      | 0
-        "hello/world"       | "/var/hello1/other"     | 5      | -1
-        "hello1/world"      | "/var/hello/other"      | 5      | 1
-        "hello/world/some"  | "/var/hello/other"      | 5      | 1
-        "hello/world"       | "/var/hello/world1"     | 5      | -1
-        "bbc/some"          | "/var/abc/other"        | 5      | 1
-        "/hello/world/some" | "/var/hello/other"      | 0      | -1
-        "/hello/world/some" | "/var/hello/other"      | 4      | 1
-        "dir1"              | "some/dir12"            | 5      | -1
-        "dir12"             | "some/dir1"             | 5      | 1
-    }
-
-    def "equal chars are equal"() {
-        expect:
-        (Character.MIN_VALUE..Character.MAX_VALUE).each { currentChar ->
-            assert compareCharsIgnoringCase(currentChar as char, currentChar as char) == 0
-            assert compareChars(currentChar as char, currentChar as char) == 0
-            CaseSensitivity.values().each {
-                assert equalChars(currentChar as char, currentChar as char, it == CaseSensitivity.CASE_SENSITIVE)
-            }
-        }
-    }
-
-    def "path separators are equal"() {
-        def slash = '/' as char
-        def backslash = '\\' as char
-        expect:
-        compareCharsIgnoringCase(slash, backslash) == 0
-        compareCharsIgnoringCase(backslash, slash) == 0
-        compareChars(slash, backslash) == 0
-        compareChars(backslash, slash) == 0
-        CaseSensitivity.values().each {
-            def caseSensitive = it == CaseSensitivity.CASE_SENSITIVE
-            assert equalChars(slash, backslash, caseSensitive)
-            assert equalChars(backslash, slash, caseSensitive)
-        }
-    }
-
-    def "can compare path separator chars correctly (#left - #right = #result)"() {
-        def char1 = left as char
-        def char2 = right as char
-        expect:
-        compareCharsIgnoringCase(char1, char2) == result
-        compareCharsIgnoringCase(char2, char1) == -result
-        compareChars(char1, char2) == result
-        compareChars(char2, char1) == -result
-        CaseSensitivity.values().each {
-            def caseSensitive = it == CaseSensitivity.CASE_SENSITIVE
-            assert equalChars(char1, char2, caseSensitive) == (result == 0)
-            assert equalChars(char2, char1, caseSensitive) == (result == 0)
-        }
-
-        where:
-        left | right | result
-        'a'  | 'a'   | 0
-        'a'  | 'b'   | -1
-        '/'  | 'a'   | -1
-        '\\' | 'a'   | -1
-        '/'  | '/'   | 0
-        '\\' | '\\'  | 0
-        '/'  | '\\'  | 0
-        '/'  | 'A'   | -1
-        '\\' | 'A'   | -1
-    }
-
-    def "can compare lower and upper case correctly (#left - #right = #result)"() {
-        def char1 = left as char
-        def char2 = right as char
-        def caseInsensitiveResult = left.toUpperCase() == right.toUpperCase() ? 0 : result
-        expect:
-        compareCharsIgnoringCase(char1, char2) == caseInsensitiveResult
-        compareCharsIgnoringCase(char2, char1) == -caseInsensitiveResult
-        compareChars(char1, char2) == Character.compare(char1, char2)
-        compareChars(char2, char1) == Character.compare(char2, char1)
-        !equalChars(char1, char2, true)
-        equalChars(char1, char2, false) == (caseInsensitiveResult == 0)
-        !equalChars(char2, char1, true)
-        equalChars(char2, char1, false) == (caseInsensitiveResult == 0)
-
-        where:
-        left | right | result
-        'a'  | 'A'   | 32
-        'a'  | 'B'   | -1
-        'A'  | 'B'   | -1
-        'A'  | 'b'   | -1
-        'a'  | 'b'   | -1
-        'z'  | 'Z'   | 32
-        'y'  | 'z'   | -1
-        'Y'  | 'z'   | -1
-        'Y'  | 'Z'   | -1
-        'y'  | 'Z'   | -1
-    }
-
-    def "#left and #right are equal ignoring case"() {
-        char char1 = left as char
-        char char2 = right as char
-        expect:
-        equalChars(char2, char1, false)
-        equalChars(char1, char2, false)
-        compareCharsIgnoringCase(char1, char2) == 0
-        compareCharsIgnoringCase(char2, char1) == 0
-
-        where:
-        left     | right
-        '\u1e9e' | 'ß'
-        '\u03a3' | '\u03c2'
-        '\u03a3' | '\u03c3'
-    }
 }
